feat(cjs-module-analyzer)!: Update packaging for RESM/NESM bridge
BREAKING CHANGE: No longer supports direct use from CommonJS
