+++ packages/cjs-module-analyzer/NEWS.md
@@ -1,0 +1,3 @@
+# Next release
+
+- Initial version.
