+++ packages/cjs-module-analyzer/index.d.ts
@@ -1,1 +1,1 @@
-export { analyzeCommonJS } from './src/main.js';
+export { analyzeCommonJS } from './index.js';
