+++ packages/cjs-module-analyzer/package.json
@@ -14,37 +14,23 @@
     "url": "https://github.com/endojs/endo/issues"
   },
   "type": "module",
-  "parsers": {
-    "js": "mjs"
-  },
-  "main": "./dist/cjs-module-analyzer.cjs",
-  "module": "./index.js",
-  "browser": "./dist/cjs-module-analyzer.umd.js",
-  "unpkg": "./dist/cjs-module-analyzer.umd.js",
+  "main": "./index.js",
   "types": "./index.d.ts",
   "exports": {
     "./package.json": "./package.json",
-    ".": {
-      "import": "./index.js",
-      "require": "./dist/cjs-module-analyzer.cjs",
-      "browser": "./dist/cjs-module-analyzer.umd.js"
-    }
+    ".": "./index.js"
   },
   "scripts": {
-    "build": "rollup --config rollup.config.js",
-    "clean": "rm -rf dist",
-    "cover": "nyc ava",
+    "build": "exit 0",
     "lint": "yarn lint:types && yarn lint:js",
-    "lint-fix": "eslint --fix '*.js' '**/*.js' 'test*/**.js'",
-    "lint:js": "eslint '*.js' '**/*.js' 'test*/**.js'",
+    "lint-fix": "eslint --fix '**/*.js'",
+    "lint:js": "eslint '**/*.js'",
     "lint:types": "tsc --build jsconfig.json",
-    "prepublish": "yarn clean && yarn build",
     "test": "ava"
   },
   "dependencies": {},
   "devDependencies": {
-    "@rollup/plugin-commonjs": "^13.0.0",
-    "@rollup/plugin-node-resolve": "^6.1.0",
+    "@agoric/eslint-config": "^0.3.6",
     "ava": "^3.12.1",
     "babel-eslint": "^10.0.3",
     "eslint": "^7.23.0",
@@ -53,17 +39,13 @@
     "eslint-plugin-eslint-comments": "^3.1.2",
     "eslint-plugin-import": "^2.19.1",
     "eslint-plugin-prettier": "^3.1.2",
-    "nyc": "^15.1.0",
     "prettier": "^1.19.1",
-    "rollup": "1.31.0",
-    "rollup-plugin-terser": "^5.1.3",
     "typescript": "^4.0.5"
   },
   "files": [
     "LICENSE*",
-    "dist",
-    "src",
-    "types"
+    "index.d.ts",
+    "index.js"
   ],
   "publishConfig": {
     "access": "public"
